plumbing: transport, refactor and define transport operations
This change redefines the transport package to be more modular and
extensible. It redefines the `transport.Transport`, `transport.Session`,
and `transport.Commander` and `transport.Command` interfaces. It also
introduces a new `transport.Connection` that abstracts the logic of
sending and receiving git data to and from a server.

There are 2 main transports. The pack transport which is a full-duplex
command based transport that uses `transport.PackSession`. And the other
is the HTTP transport which is a half-duplex transport that uses
`transport.HTTPSession`. All of file, git, and ssh transports use the
pack transport.

The rest of the abstractions are defined in `transport.NegotiatePack`,
`transport.FetchPack`, `transport.UploadPack`, `transport.SendPack`, and
`tansport.ReceivePack`. All of these static functions are used by the
`transport.Connection` to send and receive data to and from a
`storage.Storer` interface type.

Reference: https://gist.github.com/aymanbagabas/1b5ccc265ee62dd574c62d378b4b57bb